Add PID and port parsing to OSX
This commit changes 2 things for OSX ('darwin'):
1. Make `netstat` output the PID with the `-v` argument (and parse the
PID in lib/parsers)
2. Make lib/utils.parseAddress handle OSX addresses (which have the
strange format `ip.addr.port`, e.g. `127.0.0.1.22`)

Additionally, this commit fixes a bug when setting the "limit" option:
The filters.limit function was not being passed a "stop()" function,
which caused it to throw an error when trying to call stop() after
the limit was reached.