Expose the activators in exports
The default activator is not working for me (see issue #4 -- `done()` is
called before all the output is parsed), so I need to override it. I
also can't use the default `sync` activator, because on Node v0.10.43,
which I must use, there is no child_process.spawnSync. Anyway, those
are personal problems ;) but it may be useful for some others to also
have access to the activators object?