chore: rename 'UNLICENSE' file to 'LICENSE' to comply with convention
The docs at https://docs.npmjs.com/cli/v10/configuring-npm/package-json#files imply that the license file should be called 'LICENSE' or 'LICENCE'. The problem with giving the file another name is that tools like the license-webpack-plugin can't find the license in that case.

Potentially breaking if any tools depend on the current filename